refactor: replace const for validation in registartion form to utils folder
