+++ src/pages/registration/registration-form-fourth/registration-form-fourth.tsx
@@ -2,13 +2,15 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import { z } from 'zod';
 
-import { allowedCountries, postalCodePatterns } from '../const-for-validation';
-
 import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import { cn } from '@/lib/utils';
-import { ROUTES } from '@/utils/constantes';
+import {
+  allowedCountries,
+  postalCodePatterns,
+  ROUTES,
+} from '@/utils/constantes';
 
 const formSchema = z
   .object({
